requirements.txt
amqp==2.2.1
asn1crypto==0.23.0
billiard==3.5.0.3
celery==4.1.0
certifi==2017.7.27.1
cffi==1.11.2
chardet==3.0.4
coreapi==2.3.1
coreschema==0.0.4
cryptography==2.1.4
datedelta==1.2
dj-database-url==0.4.2
Django==1.11.4
django-cors-headers==2.1.0
django-filter==1.0.4
django-json-widget==0.1.0
django-model-utils==3.0.0
django-oauth-toolkit==1.0.0
django-rest-swagger==2.1.2
djangorestframework==3.6.3
ecdsa==0.13
et-xmlfile==1.0.1
future==0.16.0
gunicorn==19.7.1
idna==2.5
itypes==1.1.0
jdcal==1.3
Jinja2==2.9.6
kombu==4.1.0
MarkupSafe==1.0
numpy==1.13.3
oauthlib==2.0.2
openapi-codec==1.3.2
openpyxl==2.4.8
pandas==0.20.3
psycopg2==2.6.2
pycparser==2.18
PyJWT==1.4.0
python-dateutil==2.6.1
pytz==2017.2
pyxero==0.9.0
redis==2.10.6
requests==2.18.3
requests-oauthlib==0.8.0
simplejson==3.11.1
six==1.10.0
slacker==0.9.42
slacker-log-handler==1.6.1
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
whitenoise==3.2
